ACS-6472 Extract node event mapping.
