+++ live-ingester/src/test/java/org/alfresco/hxi_connector/live_ingester/domain/event/IngestNewNodeEventHandlerTest.java
@@ -26,91 +26,43 @@
 
 package org.alfresco.hxi_connector.live_ingester.domain.event;
 
-import static org.junit.jupiter.api.Assertions.assertTrue;
-import static org.mockito.Mockito.mock;
+import static org.mockito.BDDMockito.given;
 import static org.mockito.Mockito.verify;
-
-import static org.alfresco.hxi_connector.live_ingester.domain.model.out.PredefinedNodeProperty.*;
-import static org.alfresco.hxi_connector.live_ingester.util.TestUtils.assertContainsSameElements;
-
-import java.util.Set;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import org.mockito.ArgumentCaptor;
-import org.mockito.Captor;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 
 import org.alfresco.hxi_connector.live_ingester.domain.model.in.IngestNewNodeEvent;
-import org.alfresco.hxi_connector.live_ingester.domain.model.in.Node;
-import org.alfresco.hxi_connector.live_ingester.domain.model.out.NodeProperty;
 import org.alfresco.hxi_connector.live_ingester.domain.model.out.event.EventPublisher;
 import org.alfresco.hxi_connector.live_ingester.domain.model.out.event.UpdateNodeMetadataEvent;
 
 @ExtendWith(MockitoExtension.class)
 class IngestNewNodeEventHandlerTest
 {
-    private static final long EVENT_TIMESTAMP = 1_690_000_000_100L;
-    private static final String NODE_ID = "0fe2919a-e0a6-4033-8d35-168a16cf33fc";
-    private static final String NODE_NAME = "test-name";
-    private static final String NODE_PRIMARY_ASSOC_Q_NAME = "cm:test-name";
-    private static final String NODE_TYPE = "cm:folder";
-    private static final String NODE_CREATED_BY_USER_WITH_ID = "admin";
-    private static final String NODE_MODIFIED_BY_USER_WITH_ID = "admin";
-    private static final Set<String> NODE_ASPECT_NAMES = Set.of(
-            "cm:titled",
-            "cm:auditable");
-    private static final boolean NODE_IS_FOLDER = true;
-    private static final boolean NODE_IS_FILE = false;
-    private static final long NODE_CREATED_AT = 1_690_000_000_050L;
-    private static final NodeProperty<String> NODE_TITLE = new NodeProperty<>("cm:title", "some title");
-    private static final Set<NodeProperty<?>> NODE_PROPERTIES = Set.of(NODE_TITLE);
-    @Captor
-    ArgumentCaptor<UpdateNodeMetadataEvent> updateNodeMetadataEventCaptor;
-    private final EventPublisher eventPublisher = mock();
-    private final IngestNewNodeEventHandler ingestNewNodeEventHandler = new IngestNewNodeEventHandler(eventPublisher);
+    @Mock
+    IngestNewNodeEvent ingestNewNodeEvent;
+    @Mock
+    UpdateNodeMetadataEvent updateNodeMetadataEvent;
+    @Mock
+    UpdateNodeEventMapper updateNodeEventMapper;
+    @Mock
+    EventPublisher eventPublisher;
+    @InjectMocks
+    IngestNewNodeEventHandler ingestNewNodeEventHandler;
 
     @Test
-    void shouldSetNewlyCreatedNodeMetadataProperties()
+    void nodeWithoutContent_PublishMetadata()
     {
         // given
-        Node node = new Node(
-                NODE_ID,
-                NODE_NAME,
-                NODE_PRIMARY_ASSOC_Q_NAME,
-                NODE_TYPE,
-                NODE_CREATED_BY_USER_WITH_ID,
-                NODE_MODIFIED_BY_USER_WITH_ID,
-                NODE_ASPECT_NAMES,
-                NODE_IS_FILE,
-                NODE_IS_FOLDER,
-                NODE_CREATED_AT,
-                NODE_PROPERTIES);
-
-        IngestNewNodeEvent event = new IngestNewNodeEvent(
-                EVENT_TIMESTAMP,
-                node);
-
-        Set<NodeProperty<?>> expectedNodePropertiesToSet = Set.of(
-                NAME.withValue(NODE_NAME),
-                PRIMARY_ASSOC_Q_NAME.withValue(NODE_PRIMARY_ASSOC_Q_NAME),
-                TYPE.withValue(NODE_TYPE),
-                CREATED_BY_USER_WITH_ID.withValue(NODE_CREATED_BY_USER_WITH_ID),
-                MODIFIED_BY_USER_WITH_ID.withValue(NODE_MODIFIED_BY_USER_WITH_ID),
-                ASPECTS_NAMES.withValue(NODE_ASPECT_NAMES),
-                IS_FILE.withValue(NODE_IS_FILE),
-                IS_FOLDER.withValue(NODE_IS_FOLDER),
-                CREATED_AT.withValue(NODE_CREATED_AT),
-                NODE_TITLE);
+        given(updateNodeEventMapper.map(ingestNewNodeEvent)).willReturn(updateNodeMetadataEvent);
 
         // when
-        ingestNewNodeEventHandler.handle(event);
+        ingestNewNodeEventHandler.handle(ingestNewNodeEvent);
 
         // then
-        verify(eventPublisher).publishMessage(updateNodeMetadataEventCaptor.capture());
-        UpdateNodeMetadataEvent updateNodeMetadataEvent = updateNodeMetadataEventCaptor.getValue();
-
-        assertContainsSameElements(expectedNodePropertiesToSet, updateNodeMetadataEvent.getMetadataPropertiesToSet().values());
-        assertTrue(updateNodeMetadataEvent.getMetadataPropertiesToUnset().isEmpty(), "There should be no properties to unset");
+        verify(eventPublisher).publishMessage(updateNodeMetadataEvent);
     }
 }
